+++ Konsole.Tests/ProgressBarSlimTests/RefreshShould.cs
@@ -11,13 +11,13 @@ namespace Konsole.Tests.ProgressBarSlimTests
     {
         ////Item {0,-5} of {1,-5}. ({2,-3}%) 
         [Test]
-        [TestCase(100,0,   "1234567890 (0  %)          ", "                           ", PbStyle.SingleLine)]
-        [TestCase(101,50,  "1234567890 (50 %) ####     ", "                           ", PbStyle.SingleLine)]
-        [TestCase(102,100, "1234567890 (100%) #########", "                           ", PbStyle.SingleLine)]
-        //[TestCase(103, 0,  "Item     0 of   100. (  0%)", "                           ", PbStyle.DoubeLine)]
-        //[TestCase(104, 50, "Item    50 of   100. ( 50%)", "                           ", PbStyle.DoubeLine)]
-        //[TestCase(105, 100,"Item   100 of   100. (100%)", "                           ", PbStyle.DoubeLine)]
-        public void at_xx_pc_the_progress_bar_should_fill_the_balance_of_the_console_width_pro_ratio(int seq, int i, string line1, string line2, PbStyle style)
+        [TestCase(100,0,   "1234567890 (0  %)          ", "                           ")]
+        [TestCase(101,50,  "1234567890 (50 %) ####     ", "                           ")]
+        [TestCase(102,100, "1234567890 (100%) #########", "                           ")]
+        //[TestCase(103, 0,  "Item     0 of   100. (  0%)", "                           ")]
+        //[TestCase(104, 50, "Item    50 of   100. ( 50%)", "                           ")]
+        //[TestCase(105, 100,"Item   100 of   100. (100%)", "                           ")]
+        public void at_xx_pc_the_progress_bar_should_fill_the_balance_of_the_console_width_pro_ratio(int seq, int i, string line1, string line2)
         {
             var console = new MockConsole(27,2);
             var pb = new ProgressBar(console, 100, 10);